Put create post functionality to Posts page instead of separate page. Added basic grid layout for posts with bootstrap gridding.
