+++ app/javascript/App.jsx
@@ -2,7 +2,6 @@ import React from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom"
 
 import Posts from "./components/Posts"
-import CreatePost from "./components/CreatePost"
 
 const App = () => {
 
@@ -10,7 +9,6 @@ const App = () => {
         <Router>
             <Routes>
                 <Route path='/' exact element={<Posts/>} />
-                <Route path='/create-post' exact element={<CreatePost />} />
             </Routes>
         </Router>
     );
